refactor(api): 重构 LTM 检索架构，拆分并对称化检索器模块
- 将 LLM 语义检索从 long_term.py 分离至独立 llm_retriever.py
- 重命名 retriever.py → mechanical_retriever.py，与 llm_retriever.py 对称
- 将 RetrievalMode 枚举迁移至 long_term.py（LTM 核心文件）
- 生命周期方法标准化：start_listening/stop_listening/is_listening → start/stop/running
- MechanicalRetriever 改为 eager 初始化，与 LLMRetriever 一致
