+++ api/memory/__init__.py
@@ -1,8 +1,7 @@
 """记忆模块 — 长期记忆叙事、短期记忆检查点、用户初始化与回调。"""
 
 from api.memory.callback import MemoryToolCallback
-from api.memory.long_term import MEMORY_PATH, LongTermMemory, get_narrative
-from api.memory.retriever import RetrievalMode
+from api.memory.long_term import MEMORY_PATH, LongTermMemory, RetrievalMode, get_narrative
 from api.memory.short_term import delete_thread, get_checkpointer
 from api.memory.user_init import ensure_all, ensure_env_file, ensure_soul_md, ensure_user_md
